(Refactoring) Reorganize the project structure
