+++ data/textures/explode.png
