+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "Game.h"
+#include "game/Game.h"
 
 int main() {
     Game game;
