+++ src/managers/ResourceManager.cpp
