+++ src/managers/ResourceManager.h
